AVS-52 StreamVideo is not initialised if notification starts dogfooding app
